+++ abis/V2Factory.json
@@ -1,5 +1,123 @@
 [
 	{
+		"inputs": [
+			{
+				"internalType": "address",
+				"name": "_minionSummoner",
+				"type": "address"
+			}
+		],
+		"payable": false,
+		"stateMutability": "nonpayable",
+		"type": "constructor"
+	},
+	{
+		"anonymous": false,
+		"inputs": [
+			{
+				"indexed": true,
+				"internalType": "address",
+				"name": "moloch",
+				"type": "address"
+			},
+			{
+				"indexed": true,
+				"internalType": "address[]",
+				"name": "summoners",
+				"type": "address[]"
+			},
+			{
+				"indexed": true,
+				"internalType": "address",
+				"name": "depositToken",
+				"type": "address"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningTime",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "periodDuration",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "votingPeriodLength",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "gracePeriodLength",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "proposalDeposit",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "dilutionBound",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "processingReward",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256[]",
+				"name": "summonerStake",
+				"type": "uint256[]"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningDeposit",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningRate",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningTermination",
+				"type": "uint256"
+			}
+		],
+		"name": "SummonMoloch",
+		"type": "event"
+	},
+	{
+		"constant": true,
+		"inputs": [],
+		"name": "minionSummoner",
+		"outputs": [
+			{
+				"internalType": "contract ISummonMinion",
+				"name": "",
+				"type": "address"
+			}
+		],
+		"payable": false,
+		"stateMutability": "view",
+		"type": "function"
+	},
+	{
 		"constant": false,
 		"inputs": [
 			{
@@ -8,9 +126,9 @@
 				"type": "address[]"
 			},
 			{
-				"internalType": "address[]",
-				"name": "_approvedTokens",
-				"type": "address[]"
+				"internalType": "address",
+				"name": "_depositToken",
+				"type": "address"
 			},
 			{
 				"internalType": "uint256",
@@ -43,9 +161,9 @@
 				"type": "uint256"
 			},
 			{
-				"internalType": "uint256",
-				"name": "_summonerStake",
-				"type": "uint256"
+				"internalType": "uint256[]",
+				"name": "_summonerShares",
+				"type": "uint256[]"
 			},
 			{
 				"internalType": "uint256",
@@ -67,46 +185,6 @@
 		"outputs": [],
 		"payable": false,
 		"stateMutability": "nonpayable",
-		"type": "function"
-	},
-	{
-		"inputs": [
-			{
-				"internalType": "address",
-				"name": "_minionSummoner",
-				"type": "address"
-			}
-		],
-		"payable": false,
-		"stateMutability": "nonpayable",
-		"type": "constructor"
-	},
-	{
-		"anonymous": false,
-		"inputs": [
-			{
-				"indexed": true,
-				"internalType": "address",
-				"name": "moloch",
-				"type": "address"
-			}
-		],
-		"name": "SummonMoloch",
-		"type": "event"
-	},
-	{
-		"constant": true,
-		"inputs": [],
-		"name": "minionSummoner",
-		"outputs": [
-			{
-				"internalType": "contract ISummonMinion",
-				"name": "",
-				"type": "address"
-			}
-		],
-		"payable": false,
-		"stateMutability": "view",
 		"type": "function"
 	}
 ]
