+++ src/factory-mapping.ts
@@ -3,36 +3,55 @@ import { SummonMoloch } from "../generated/MolochSummoner/V2Factory";
 
 import { MolochTemplate } from "../generated/templates";
 import { Moloch } from "../generated/schema";
+import { createAndApproveToken, createEscrowTokenBalance, createGuildTokenBalance, createAndAddSummoner} from "./v2-mapping"
 
 
 export function handleSummoned(event: SummonMoloch): void {
   
-  let entity = Moloch.load(event.params.moloch.toHex())
-
-  // Entities only exist after they have been saved to the store;
-  // `null` checks allow to create entities on demand
-  if (entity == null) {
-    entity = new Moloch(event.params.moloch.toHex())
-  }
-
   MolochTemplate.create(event.params.moloch);
+
   let molochId = event.params.moloch.toHex();
   let moloch = new Moloch(molochId);
+  let tokens = event.params.tokens;
+  let approvedTokens: string[] = [];
 
-  moloch.summoningTime = BigInt.fromI32(1);
-  moloch.version = "2x";
-  moloch.periodDuration = BigInt.fromI32(1);
-  moloch.votingPeriodLength = BigInt.fromI32(1);
-  moloch.gracePeriodLength = BigInt.fromI32(1);
-  moloch.proposalDeposit = BigInt.fromI32(1);
-  moloch.dilutionBound = BigInt.fromI32(3);
-  moloch.processingReward = BigInt.fromI32(1);
-  moloch.approvedTokens = new Array<string>();
-  moloch.guildTokenBalance = new Array<string>();
-  moloch.escrowTokenBalance = new Array<string>();
-  moloch.totalShares = BigInt.fromI32(0);
-  moloch.summoningRate = BigInt.fromI32(1);
-  moloch.summoningTermination = BigInt.fromI32(1);
+  let escrowTokenBalance: string[] = [];
+  let guildTokenBalance: string[] = [];
+
+  for (let i = 0; i < tokens.length; i++) {
+    let token = tokens[i];
+    approvedTokens.push(createAndApproveToken(molochId, token));
+    escrowTokenBalance.push(createEscrowTokenBalance(molochId, token));
+    guildTokenBalance.push(createGuildTokenBalance(molochId, token));
+  }
+
+  let eventSummoners = event.params.summoners;
+  let summoners: string[] = [];
+
+  for (let i = 0; i < eventSummoners.length; i++) {
+    let summoner = eventSummoners[i];
+    summoners.push(
+      createAndAddSummoner(molochId, summoner as Bytes, tokens, event)
+    );
+  }
+
+  moloch.summoner = event.params.summoner;
+  moloch.summoningTime = event.params._summoningTime;
+  moloch.title = event.params.title;
+  moloch.version = "2";
+  moloch.deleted = false;
+  moloch.newContract = "1";
+  moloch.periodDuration = event.params._periodDuration;
+  moloch.votingPeriodLength = event.params._votingPeriodLength;
+  moloch.gracePeriodLength = event.params._gracePeriodLength;
+  moloch.proposalDeposit = event.params._proposalDeposit;
+  moloch.dilutionBound = event.params._dilutionBound;
+  moloch.processingReward = event.params._processingReward;
+  moloch.depositToken = approvedTokens[0];
+  moloch.approvedTokens = approvedTokens;
+  moloch.guildTokenBalance = guildTokenBalance;
+  moloch.escrowTokenBalance = escrowTokenBalance;
+  moloch.totalShares = BigInt.fromI32(1);
   moloch.totalLoot = BigInt.fromI32(0);
   moloch.proposalCount = BigInt.fromI32(0);
   moloch.proposalQueueCount = BigInt.fromI32(0);
@@ -41,7 +60,7 @@ export function handleSummoned(event: SummonMoloch): void {
   moloch.proposedToKick = new Array<string>();
   moloch.proposedToFund = new Array<string>();
   moloch.proposedToTrade = new Array<string>();
-  
+
   moloch.save();
